+++ MANIFEST
@@ -1,0 +1,4 @@
+# file GENERATED by distutils, do NOT edit
+setup.cfg
+setup.py
+wekapy/wekapy.py
