+++ setup.py
@@ -1,0 +1,14 @@
+from distutils.core import setup
+
+setup(
+    name = 'wekapy',
+    packages = ['wekapy'],
+    version = '1.3.1',
+    description = 'Simple Python wrapper for the WEKA toolkit.',
+    maintainer = 'Faiz Siddiqui',
+    maintainer_email = 'faize.sid@gmail.com',
+    url = 'https://github.com/faizsiddiqui/WekaPy',
+    download_url = 'https://github.com/faizsiddiqui/WekaPy/archive/1.3.1.tar.gz',
+    keywords = ['WEKA', 'Machine Learning', 'Data Mining', 'Python'],
+    classifiers = [],
+)
