+++ wekapy/WekapyException.py
@@ -1,0 +1,6 @@
+class WekapyException(Exception):
+    def __init__(self, message):
+        self.message = message
+
+    def __str__(self):
+        return self.message
