+++ wekapy/__init__.py
@@ -1,0 +1,20 @@
+# Portions of this software Copyright 2017 Faiz Siddiqui
+# Portions of this software Copyright 2013 Will Webberley
+# Portions of this software Copyright 2013 Martin Chorley
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# The full license text is available at <http://www.gnu.org/licenses/>.
+
+from Model import Model
+from Filter import Filter
+from Feature import Feature
+from Instance import Instance
